+++ tests/test_flask.py
@@ -1,5 +1,8 @@
+# encoding=utf8
 import unittest
 from flask import Flask, json, abort
+from flask.ext import restful
+
 from wsmeext.flask import signature
 from wsme.api import Response
 from wsme.types import Base, text
@@ -16,6 +19,7 @@ class Criterion(Base):
     value = text
 
 test_app = Flask(__name__)
+api = restful.Api(test_app)
 
 
 @test_app.route('/multiply')
@@ -78,6 +82,18 @@ def get_status_response():
     return Response(1, status_code=201)
 
 
+class RestFullApi(restful.Resource):
+    @signature(Model)
+    def get(self):
+        return Model(id=1, name=u"Gérard")
+
+    @signature(int, body=Model)
+    def post(self, model):
+        return model.id
+
+api.add_resource(RestFullApi, '/restful')
+
+
 class FlaskrTestCase(unittest.TestCase):
 
     def setUp(self):
@@ -89,7 +105,7 @@ class FlaskrTestCase(unittest.TestCase):
 
     def test_multiply(self):
         r = self.app.get('/multiply?a=2&b=5')
-        assert r.data == '10'
+        assert r.data == b'10', r.data
 
     def test_get_model(self):
         resp = self.app.get('/models/test')
@@ -102,9 +118,8 @@ class FlaskrTestCase(unittest.TestCase):
     def test_array_parameter(self):
         resp = self.app.get('/models?q.op=%3D&q.attr=name&q.value=second')
         assert resp.status_code == 200
-        print resp.data
-        self.assertEquals(
-            resp.data, '[{"name": "second"}]'
+        self.assertEqual(
+            resp.data, b'[{"name": "second"}]'
         )
 
     def test_post_model(self):
@@ -138,9 +153,9 @@ class FlaskrTestCase(unittest.TestCase):
             headers={'Accept': 'application/xml'}
         )
         assert r.status_code == 403, r.status_code
-        assert r.data == ('<error><faultcode>Client</faultcode>'
-                          '<faultstring>403: Forbidden</faultstring>'
-                          '<debuginfo /></error>')
+        assert r.data == (b'<error><faultcode>Client</faultcode>'
+                          b'<faultstring>403: Forbidden</faultstring>'
+                          b'<debuginfo /></error>')
 
     def test_custom_non_http_clientside_error(self):
         r = self.app.get(
@@ -155,18 +170,39 @@ class FlaskrTestCase(unittest.TestCase):
             headers={'Accept': 'application/xml'}
         )
         assert r.status_code == 412, r.status_code
-        assert r.data == ('<error><faultcode>Client</faultcode>'
-                          '<faultstring>FOO!</faultstring>'
-                          '<debuginfo /></error>')
+        assert r.data == (b'<error><faultcode>Client</faultcode>'
+                          b'<faultstring>FOO!</faultstring>'
+                          b'<debuginfo /></error>')
 
     def test_serversideerror(self):
         r = self.app.get('/divide_by_zero')
         assert r.status_code == 500
-        self.assertEquals(
-            r.data,
-            '{"debuginfo": null, "faultcode": "Server", "faultstring": '
-            '"integer division or modulo by zero"}'
-        )
+        data = json.loads(r.data)
+        self.assertEqual(data['debuginfo'], None)
+        self.assertEqual(data['faultcode'], 'Server')
+        self.assertIn('by zero', data['faultstring'])
+
+    def test_restful_get(self):
+        r = self.app.get('/restful', headers={'Accept': 'application/json'})
+        self.assertEqual(r.status_code, 200)
+
+        data = json.loads(r.data)
+
+        self.assertEqual(data['id'], 1)
+        self.assertEqual(data['name'], u"Gérard")
+
+    def test_restful_post(self):
+        r = self.app.post(
+            '/restful',
+            data=json.dumps({'id': 5, 'name': u'Huguette'}),
+            headers={
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'})
+        self.assertEqual(r.status_code, 200)
+
+        data = json.loads(r.data)
+
+        self.assertEqual(data, 5)
 
 if __name__ == '__main__':
     test_app.run()