+++ wsme/tests/test_protocols_commons.py
@@ -3,13 +3,19 @@
 import datetime
 import unittest
 
-from wsme.rest.args import from_param, from_params
+from wsme.api import FunctionArgument, FunctionDefinition
+from wsme.rest.args import from_param, from_params, args_from_args
+from wsme.exc import InvalidInput
 
 from wsme.types import UserType, Unset, ArrayType, DictType
 
 
 class MyUserType(UserType):
     basetype = str
+
+
+class DictBasedUserType(UserType):
+    basetype = DictType(int, int)
 
 
 class TestProtocolsCommons(unittest.TestCase):
@@ -52,3 +58,62 @@ class TestProtocolsCommons(unittest.TestCase):
 
     def test_from_params_dict_unset(self):
         assert from_params(DictType(int, str), {}, 'a', set()) is Unset
+
+    def test_from_params_usertype(self):
+        value = from_params(
+            DictBasedUserType(),
+            {'a[2]': '2'},
+            'a',
+            set()
+        )
+        self.assertEqual(value, {2: 2})
+
+    def test_args_from_args_usertype(self):
+
+        class FakeType(UserType):
+            name = 'fake-type'
+            basetype = int
+
+        fake_type = FakeType()
+        fd = FunctionDefinition(FunctionDefinition)
+        fd.arguments.append(FunctionArgument('fake-arg', fake_type, True, 0))
+
+        new_args = args_from_args(fd, [1], {})
+        self.assertEqual([1], new_args[0])
+
+        # can't convert str to int
+        try:
+            args_from_args(fd, ['invalid-argument'], {})
+        except InvalidInput as e:
+            assert fake_type.name in str(e)
+        else:
+            self.fail('Should have thrown an InvalidInput')
+
+
+class ArgTypeConversion(unittest.TestCase):
+
+    def test_int_zero(self):
+        self.assertEqual(0, from_param(int, 0))
+        self.assertEqual(0, from_param(int, '0'))
+
+    def test_int_nonzero(self):
+        self.assertEqual(1, from_param(int, 1))
+        self.assertEqual(1, from_param(int, '1'))
+
+    def test_int_none(self):
+        self.assertEqual(None, from_param(int, None))
+
+    def test_float_zero(self):
+        self.assertEqual(0.0, from_param(float, 0))
+        self.assertEqual(0.0, from_param(float, 0.0))
+        self.assertEqual(0.0, from_param(float, '0'))
+        self.assertEqual(0.0, from_param(float, '0.0'))
+
+    def test_float_nonzero(self):
+        self.assertEqual(1.0, from_param(float, 1))
+        self.assertEqual(1.0, from_param(float, 1.0))
+        self.assertEqual(1.0, from_param(float, '1'))
+        self.assertEqual(1.0, from_param(float, '1.0'))
+
+    def test_float_none(self):
+        self.assertEqual(None, from_param(float, None))