+++ wsme/tests/test_types.py
@@ -326,27 +326,44 @@ Value: 'v3'. Value should be one of: v., v.",
         v.validate('A')
         self.assertRaises(ValueError, v.validate, '_')
 
+    def test_validate_string_type_pattern_exception_message(self):
+        regex = '^[a-zA-Z0-9]*$'
+        v = types.StringType(pattern=regex)
+        try:
+            v.validate('_')
+            self.assertFail()
+        except ValueError as e:
+            self.assertIn(regex, str(e))
+
     def test_validate_ipv4_address_type(self):
         v = types.IPv4AddressType()
-        v.validate('127.0.0.1')
-        v.validate('192.168.0.1')
+        self.assertEqual(v.validate('127.0.0.1'), '127.0.0.1')
+        self.assertEqual(v.validate('192.168.0.1'), '192.168.0.1')
+        self.assertEqual(v.validate(u'8.8.1.1'), u'8.8.1.1')
         self.assertRaises(ValueError, v.validate, '')
         self.assertRaises(ValueError, v.validate, 'foo')
         self.assertRaises(ValueError, v.validate,
                           '2001:0db8:bd05:01d2:288a:1fc0:0001:10ee')
+        self.assertRaises(ValueError, v.validate, '1.2.3')
 
     def test_validate_ipv6_address_type(self):
         v = types.IPv6AddressType()
-        v.validate('0:0:0:0:0:0:0:1')
-        v.validate('2001:0db8:bd05:01d2:288a:1fc0:0001:10ee')
+        self.assertEqual(v.validate('0:0:0:0:0:0:0:1'),
+                         '0:0:0:0:0:0:0:1')
+        self.assertEqual(v.validate(u'0:0:0:0:0:0:0:1'), u'0:0:0:0:0:0:0:1')
+        self.assertEqual(v.validate('2001:0db8:bd05:01d2:288a:1fc0:0001:10ee'),
+                         '2001:0db8:bd05:01d2:288a:1fc0:0001:10ee')
         self.assertRaises(ValueError, v.validate, '')
         self.assertRaises(ValueError, v.validate, 'foo')
         self.assertRaises(ValueError, v.validate, '192.168.0.1')
+        self.assertRaises(ValueError, v.validate, '0:0:0:0:0:0:1')
 
     def test_validate_uuid_type(self):
         v = types.UuidType()
-        v.validate('6a0a707c-45ef-4758-b533-e55adddba8ce')
-        v.validate('6a0a707c45ef4758b533e55adddba8ce')
+        self.assertEqual(v.validate('6a0a707c-45ef-4758-b533-e55adddba8ce'),
+                         '6a0a707c-45ef-4758-b533-e55adddba8ce')
+        self.assertEqual(v.validate('6a0a707c45ef4758b533e55adddba8ce'),
+                         '6a0a707c-45ef-4758-b533-e55adddba8ce')
         self.assertRaises(ValueError, v.validate, '')
         self.assertRaises(ValueError, v.validate, 'foo')
         self.assertRaises(ValueError, v.validate,
@@ -627,3 +644,11 @@ Value: 'v3'. Value should be one of: v., v.",
         TempType.add_attributes(two=int)
         after = types.list_attributes(TempType)
         self.assertEqual(len(after), 2)
+
+    def test_non_registered_complex_type(self):
+        class TempType(types.Base):
+            __registry__ = None
+
+        self.assertFalse(types.iscomplex(TempType))
+        types.registry.register(TempType)
+        self.assertTrue(types.iscomplex(TempType))